Change all

+++ image_converter.py
@@ -13,4 +13,5 @@ def convert_images(directory):
             rgb_img = img.convert('RGB')
             rgb_img = rgb_img.resize((256, 256))
             rgb_img.save(os.path.join(directory, '{}.jpg'.format(just_name)))
-            os.remove(os.path.join(directory, file_name))
+            if not file_name.endswith('.jpg'):
+                os.remove(os.path.join(directory, file_name))
